database connection manager split
