+++ Assets/MyScripts/Plan/DatabaseConnectionManager.cs
@@ -9,12 +9,33 @@ namespace U1
     {
         [SerializeField] string loadDataOnLogInURL;
         [SerializeField] string updateTaskStatusesURL;
-        [SerializeField] string saveMavLevelURL;
+        [SerializeField] string saveMaxLevelURL;
         [SerializeField] string updatePOURL;
         [SerializeField] string buyPOURL;
         private int myPlayerID;
+        public void SetPlayerID(int toSet)
+        {
+            myPlayerID = toSet;
+        }
+        public int GetPlayerID()
+        {
+            return myPlayerID;
+        }
         private SceneStartManager startManager;
-
+        public string BoolToString(bool toConvert)
+        {
+            if (toConvert)
+                return "1";
+            else
+                return "0";
+        }
+        public bool StringToBool(string toConvert)
+        {
+            if (toConvert == "1")
+                return true;
+            else
+                return false;
+        }
         private void OnEnable()
         {
             startManager = GetComponent<SceneStartManager>();
@@ -35,189 +56,29 @@ namespace U1
 
         private void LaunchLoadDataOnLogIn(string usernameToPass)
         {
-            StartCoroutine(LoadDataOnLogIn(usernameToPass));
+            string[] dataToPass = new string[2];
+            dataToPass[0] = usernameToPass;
+            dataToPass[1] = loadDataOnLogInURL;
+            DCMLoadOnLogin LoadDataOnLogIn = new DCMLoadOnLogin(this, this, startManager, dataToPass);
         }
-        IEnumerator LoadDataOnLogIn(string usernameToPass)
-        {
-            WWWForm wFrom = new WWWForm();
-            wFrom.AddField("username", usernameToPass);
-            using (UnityWebRequest webRequest = UnityWebRequest.Post(loadDataOnLogInURL, wFrom))
-            {
-                yield return webRequest.SendWebRequest();
-                if (webRequest.isNetworkError || webRequest.isHttpError)
-                {
-                    Debug.Log(": Error: " + webRequest.error);
-                }
-                else if(webRequest.downloadHandler.text == "0")
-                {
-                    Debug.Log("Sth went wrong with php: " + webRequest.error);
-                }
-                else
-                {
-                    //Debug.Log("Data recieved: " + webRequest.downloadHandler.text);
-                    UpdateDataOnLogIn(webRequest.downloadHandler.text);
-                }
-            }
-        }
-        private void UpdateDataOnLogIn(string webData)
-        {
-            //Debug.Log(webData);
-            string[] firstDataSplit = webData.Split('^');
-
-            string[] playerIDAndLevel = firstDataSplit[0].Split('/');
-            try
-            {
-                int playerId = Int32.Parse(playerIDAndLevel[0]);
-                int playerMaxLevel = Int32.Parse(playerIDAndLevel[1]);
-                int playerCoins = Int32.Parse(playerIDAndLevel[2]);
-                int playerExperience = Int32.Parse(playerIDAndLevel[3]);
-                int playerMaxSlots = Int32.Parse(playerIDAndLevel[4]);
-                myPlayerID = playerId;
-                startManager.SetMaxAllowLevel(playerMaxLevel);
-                startManager.SetPlayerCoins(playerCoins);
-                startManager.SetPlayerExperience(playerExperience);
-                startManager.SetPlayerMaxSlots(playerMaxSlots);
-                //Debug.Log("Id sucessfully converted to: " + playerId);
-            }
-            catch
-            {
-                Debug.Log("Id failed to parse");
-            }
-            string[] allTasksInfo = firstDataSplit[1].Split('|');
-            int tasksInfoLength = allTasksInfo.Length;
-            for (int i = 0; i < tasksInfoLength; i++)
-            {
-                string[] oneTaskInfo = allTasksInfo[i].Split('/');
-                //Debug.Log("Index I  = " + oneTaskInfo[0] + " index J = " + oneTaskInfo[1] + " value = " + oneTaskInfo[2]);
-                try
-                {
-                    int indexOne = Int32.Parse(oneTaskInfo[0]);
-                    int indexTwo = Int32.Parse(oneTaskInfo[1]);
-                    bool valueToSet = (oneTaskInfo[2]=="1");
-                    startManager.SetTaskStatuses(indexOne, indexTwo, valueToSet);
-                }
-                catch
-                {
-                    Debug.Log("Failed to parse to int");
-                }               
-            }
-            try//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-            {
-                if (firstDataSplit.Length > 2)
-                {
-                    string[] placeableObjInfo = firstDataSplit[2].Split('|');
-                    PlaceableObject[] myPlaceableObjects = startManager.GetPlaceableObjects();
-                    int objInfoLength = placeableObjInfo.Length;
-                    int placeableObjLength = myPlaceableObjects.Length;
-                    for (int i = 0; i < objInfoLength; i++)
-                    {
-                        string[] singleObjectInfo = placeableObjInfo[i].Split('/');
-                        int objIndex = Int32.Parse(singleObjectInfo[0]);
-                        for (int j = 0; j < placeableObjLength; j++)
-                        {
-                            //Debug.Log("Looping po");
-                            if (objIndex == j)
-                            {
-                                //Debug.Log("matchnig data for given index: " + objIndex + " po: " + j);
-                                myPlaceableObjects[j].numOfOwnedObjects = Int32.Parse(singleObjectInfo[1]);
-                                myPlaceableObjects[j].maxNumOfOwnedObjects = Int32.Parse(singleObjectInfo[2]);
-                                myPlaceableObjects[j].numOfObjOnStack = Int32.Parse(singleObjectInfo[3]);
-                                myPlaceableObjects[j].isAvailable = StringToBool(singleObjectInfo[4]);
-                                myPlaceableObjects[j].isAddedToStack = StringToBool(singleObjectInfo[5]);
-                                break;
-                            }
-                        }
-                    }
-                    startManager.SetPlaceableObjects(myPlaceableObjects);
-                }
-                else
-                {
-                    Debug.Log("no objects data");
-                }
-            }
-            catch
-            {
-                Debug.Log("Placeable process messed up");
-            }///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        }
-
         private void LaunchUpdateTaskStatuses(string dummy)
         {
-            StartCoroutine(UpdateTaskStatuses());
+            DCMUpdateTaskStatuses UpdateTaskStatuses = new DCMUpdateTaskStatuses(this, this, startManager, updateTaskStatusesURL);
         }
-        IEnumerator UpdateTaskStatuses()
-        {
-            WWWForm wFrom = new WWWForm();
-            wFrom.AddField("player_id", myPlayerID.ToString());
-            wFrom.AddField("task_status", CreateNewTaskStatuses());
-            using (UnityWebRequest webRequest = UnityWebRequest.Post(updateTaskStatusesURL, wFrom))
-            {
-                yield return webRequest.SendWebRequest();
-                if (webRequest.isNetworkError || webRequest.isHttpError)
-                {
-                    Debug.Log(": Error: " + webRequest.error);
-                }
-                else if (webRequest.downloadHandler.text == "0")
-                {
-                    Debug.Log("Sth went wrong with php: " + webRequest.error);
-                }
-                else if (webRequest.downloadHandler.text == "1")
-                {
-                    Debug.Log("Sucessfull Task Update  ");
-                }
-                else
-                {
-                    Debug.Log("Error:  " + webRequest.downloadHandler.text);
-                }
-            }
-        }
-        private string CreateNewTaskStatuses()
-        {
-            string taskInfoToPass = "";
-            int firstDimmLength = startManager.GetTaskStatuses().GetLength(0);
-            int secondDimmLength = startManager.GetTaskStatuses().GetLength(1);
-            for (int i = 0; i < firstDimmLength; i++)
-            {
-                for (int j = 0; j < secondDimmLength; j++)
-                {
-                    taskInfoToPass += (i.ToString() + '/' + j.ToString() + '/' + BoolToString(startManager.GetTaskStatuses()[i, j]));
-                    if (j == secondDimmLength - 1 && i == firstDimmLength - 1)
-                    {
-                        
-                    }
-                    else
-                    {
-                        taskInfoToPass += '|';
-                    }
-                }
-            }
-            return taskInfoToPass;
-        }
-        private string BoolToString(bool toConvert)
-        {
-            if (toConvert)
-                return "1";
-            else
-                return "0";
-        }
-        private bool StringToBool(string toConvert)
-        {
-            if (toConvert == "1")
-                return true;
-            else
-                return false;
-        }
-
         private void LaunchSaveMaxLevel(string maxLevelToSet)
         {
-            StartCoroutine(SaveMaxLevel(maxLevelToSet));
+            string[] dataToPass = new string[2];
+            dataToPass[0] = maxLevelToSet;
+            dataToPass[1] = saveMaxLevelURL;
+            DCMSaveMaxLevel SaveMaxLevel = new DCMSaveMaxLevel(this, this, dataToPass);
+            //StartCoroutine(SaveMaxLevel(maxLevelToSet));
         }
         IEnumerator SaveMaxLevel(string maxLevelToSet)
         {
             WWWForm wFrom = new WWWForm();
             wFrom.AddField("player_id", myPlayerID.ToString());
             wFrom.AddField("player_max_level", maxLevelToSet);
-            using (UnityWebRequest webRequest = UnityWebRequest.Post(saveMavLevelURL, wFrom))
+            using (UnityWebRequest webRequest = UnityWebRequest.Post(saveMaxLevelURL, wFrom))
             {
                 yield return webRequest.SendWebRequest();
                 if (webRequest.isNetworkError || webRequest.isHttpError)
