--- Assets/MyScripts/Plan/Plan_Manager.cs
@@ -1,1 +1,0 @@
-﻿
